+++ framework/graphics/types.h
@@ -191,6 +191,6 @@ struct Mesh_Settings {
 
 enum Data_Type data_type_get_element_type(enum Data_Type value);
 uint32_t data_type_get_count(enum Data_Type value);
-size_t data_type_get_size(enum Data_Type value);
+uint32_t data_type_get_size(enum Data_Type value);
 
 #endif
